+++ my-profile-webapp/.gitignore
@@ -1,1 +1,2 @@
 src/logs/
+/target/
