+++ my-profile-webapp/pom.xml
@@ -13,7 +13,6 @@
     <dependency>
       <groupId>edu.wisc.my.portal.angular</groupId>
       <artifactId>angularjs-portal-frame</artifactId>
-      <version>4.1.1.18</version>
       <type>war</type>
     </dependency>
     <dependency>
@@ -111,10 +110,10 @@
         <version>1.3.3</version>
         <configuration>
             <sourceDirectory>${project.basedir}/src/main/webapp/css/</sourceDirectory>
-            <outputDirectory>${project.build.directory}/${project.build.finalName}/css/</outputDirectory>
+            <outputDirectory>${project.basedir}/src/main/webapp/my-app</outputDirectory>
             <compress>true</compress>
             <includes>
-                <include>app.less</include>
+                <include>my-app.less</include>
             </includes>
             <lessJs>${project.basedir}/src/main/webapp/css/lib/less-1.6.2.js</lessJs>
         </configuration>
