+++ my-profile-webapp/src/main/webapp/my-app/.gitignore
@@ -1,0 +1,1 @@
+/my-app.css
