+++ my-profile-webapp/src/main/webapp/my-app/lec/partials/admin-lookup.html
@@ -1,4 +1,3 @@
-<link href="css/app.css" rel="stylesheet" type="text/css"/>
 <div class="portlet-frame" style='display: inline-block; width: 100%;'>
   <portlet-header app-title="Find Local Information" app-image="img/square.jpg"></portlet-header>
   <div class="portlet-body">
