+++ my-profile-webapp/src/main/webapp/my-app/lec/partials/emergency-info.html
@@ -1,4 +1,3 @@
-<link href="css/app.css" rel="stylesheet" type="text/css"/>
 <div class="portlet-frame" style='display: inline-block; width: 100%;'>
     <portlet-header app-title="My Profile - Emergency Info" app-image="img/square.jpg"></portlet-header>
