+++ my-profile-webapp/src/main/webapp/my-app/lec/partials/user.html
