+++ pom.xml
@@ -43,7 +43,13 @@
     </repositories>
     <dependencyManagement>
         <dependencies>
-        
+            <dependency>
+              <groupId>edu.wisc.my.portal.angular</groupId>
+              <artifactId>angularjs-portal-frame</artifactId>
+              <version>4.1.1.22</version>
+              <type>war</type>
+            </dependency>
+            
             <dependency>
                 <groupId>joda-time</groupId>
                 <artifactId>joda-time</artifactId>
